+++ mindspore/core/ops/angle.cc
@@ -53,7 +53,7 @@ AbstractBasePtr AngleInfer(const abstract::AnalysisEnginePtr &, const PrimitiveP
                            const std::vector<AbstractBasePtr> &input_args) {
   MS_EXCEPTION_IF_NULL(primitive);
   auto primitive_name = primitive->name();
-  const size_t input_num = 1;
+  const int64_t input_num = 1;
   CheckAndConvertUtils::CheckInputArgs(input_args, kEqual, input_num, primitive_name);
   MS_EXCEPTION_IF_NULL(input_args[0]);
   auto infer_type = AngleInferType(primitive, input_args);
